chore: apply services rules

+++ src/users/users.service.ts
@@ -1,4 +1,8 @@
-import { Body, Injectable } from '@nestjs/common';
+import {
+  ConflictException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { User, Prisma } from '@prisma/client';
 import { UsersRepository } from './users.repository';
 import { CreateUserDTO } from './dto/createUserDTO';
@@ -8,7 +12,11 @@ export class UsersService {
   constructor(private usersRepository: UsersRepository) {}
 
   async getUser(id: number) {
-    return this.usersRepository.getUser(id);
+    const user = await this.usersRepository.getUser(id);
+
+    if (!user) throw new NotFoundException('Não há usuário com esse nome');
+
+    return user;
   }
 
   async getUsers(): Promise<User[]> {
@@ -16,14 +24,25 @@ export class UsersService {
   }
 
   async createUser(data: CreateUserDTO): Promise<User> {
+    const userExists = this.usersRepository.getUserByEmail(data.email);
+
+    if (userExists) throw new ConflictException('Email já cadastrado');
     return this.usersRepository.createUser(data);
   }
 
   async updateUser(data: CreateUserDTO) {
+    const userExists = this.usersRepository.getUserByEmail(data.email);
+
+    if (!userExists) throw new NotFoundException('Usuário não encontrado');
+
     return this.usersRepository.updateUser(data);
   }
 
   async deleteUser(id: number): Promise<User> {
+    const userExists = this.usersRepository.getUser(id);
+
+    if (!userExists) throw new NotFoundException('Usuário não encontrado');
+
     return this.usersRepository.deleteUser(id);
   }
 }
